AftermintFunction/requirements.txt
requests
aws-lambda-powertools
pynamodb
pyjwt
web3